Revisa e atualiza categories
- Remove import desnecessário
- Remove métodos que não são utilizados ou versão obj-c não necessária
