--- Caronae/Categories/NSDate+isInTheFuture.swift
@@ -1,19 +1,0 @@
-import Foundation
-
-extension NSDate {
-    /// Determine if the date is in the future.
-    ///
-    /// - returns: true if date is in the future or false if now or in the past
-    @objc public func isInTheFuture() -> Bool {
-        return self.timeIntervalSinceNow > 0
-    }
-}
-
-extension Date {
-    /// Determine if the date is in the future.
-    ///
-    /// - returns: true if date is in the future or false if now or in the past
-    public func isInTheFuture() -> Bool {
-        return self.timeIntervalSinceNow > 0
-    }
-}
