--- Caronae/Categories/UIImage+imageWithTint.swift
@@ -1,29 +1,0 @@
-import UIKit
-
-extension UIImage {
-    func imageWithTintColor(_ color: UIColor) -> UIImage {
-        let rect = CGRect(origin: CGPoint.zero, size: self.size)
-        UIGraphicsBeginImageContextWithOptions(self.size, false, self.scale)
-        let context = UIGraphicsGetCurrentContext()
-        self.draw(in: rect)
-        color.setFill()
-        context?.setBlendMode(CGBlendMode.sourceAtop)
-        context?.fill(rect)
-        let result = UIGraphicsGetImageFromCurrentImageContext()
-        UIGraphicsEndImageContext()
-        return result!
-    }
-    
-    func imageWithRed(_ red: CGFloat, green: CGFloat, blue: CGFloat, alpha: CGFloat) -> UIImage {
-        let rect = CGRect(origin: CGPoint.zero, size: self.size)
-        UIGraphicsBeginImageContextWithOptions(self.size, false, self.scale)
-        let context = UIGraphicsGetCurrentContext()
-        self.draw(in: rect)
-        context?.setFillColor(red: red, green: green, blue: blue, alpha: alpha)
-        context?.setBlendMode(CGBlendMode.sourceAtop)
-        context?.fill(rect)
-        let result = UIGraphicsGetImageFromCurrentImageContext()
-        UIGraphicsEndImageContext()
-        return result!
-    }
-}
